# bank_frontend/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /withdraw:
    post:
      tags:
      - banking-controller
      operationId: withdraw
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Payment"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /transfer:
    post:
      tags:
      - banking-controller
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Payment"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /deposit:
    post:
      tags:
      - banking-controller
      operationId: deposit
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Payment"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /auth/register:
    post:
      tags:
      - auth-controller
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /auth/login:
    post:
      tags:
      - auth-controller
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /profile:
    get:
      tags:
      - banking-controller
      operationId: homePage
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: integer
                format: int64
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /mybalance:
    get:
      tags:
      - banking-controller
      operationId: getMyBalance
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: integer
                format: int64
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: number
                format: double
components:
  schemas:
    Payment:
      type: object
      properties:
        from_id:
          type: integer
          format: int64
        to_id:
          type: integer
          format: int64
        amount:
          type: number
          format: double
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ac:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        balance:
          type: number
          format: double
        version:
          type: integer
          format: int64
      required:
      - password
      - username